添加 编排者-工作者+评估者-优化者 的架构模式

+++ ai-agent-station-domain/src/main/java/top/kimwonjoon/domain/agent/service/rag/element/SimilarTextSplitter.java
@@ -164,8 +164,8 @@ public class SimilarTextSplitter {
 
         for (int i = 0; i < vectorA.length; i++) {
             dotProduct += vectorA[i] * vectorB[i];
-            normA += Math.pow(vectorA[i], 2);
-            normB += Math.pow(vectorB[i], 2);
+            normA += (float) Math.pow(vectorA[i], 2);
+            normB += (float) Math.pow(vectorB[i], 2);
         }
 
         if (normA == 0 || normB == 0) {
